src/peer-node/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1
python-multipart==0.0.6
pydantic-settings==2.1.0
